+++ pytabular/__init__.py
@@ -49,8 +49,6 @@ logger.info("Importing specifics in module...")
 from .pytabular import Tabular
 
 from .basic_checks import (
-    Return_Zero_Row_Tables,
-    Table_Last_Refresh_Times,
     BPA_Violations_To_DF,
     Last_X_Interval,
 )
